// admin/types/type.d.ts
export interface Frame {
  frame_id: string;
  name: string;
  url: string;
  created_at: string;
}